Implemented a grid of game objects for game board. Gave each game tile knowledge of its adjacent tiles.
Changed up the whole game board from a tile map to a system of 45 game objects. Each object is attached  a tile script which holds an array of tiles to indicate which tiles are adjacent to it. This enables us to attach player movement logic. I've added a player object (black circle) which moves when you click roll dice and just picked the first item in the array when there are multiple paths to choose from. It's kinda buggy and seems to stop working after roll dice is clicked a few times but it is a start.