Updated player movement
Added some logic to control how the player moves and smoothed out the token movement.
Added variable to RuleController which holds the active player (the person whose turn it is) and the dice roller consults that information to determine which playerToken object to invoke move Token on.
 Added more playerToken game objects and gave them each an offset:
Player 1 is always in the upper left of a game square
Player 2 is always in the upper right of a game square
Player 3 is always in the lower left of a game square
Player 4 is always in the lower right of a game square