Dont Destroy On Load
Created Game Object in GameSetup Scene. Attached RuleController to it. Set RuleController to not be destroyed on load.
Set the number of players selected during game setup to a value in the rule controller.
Used that to determine how many player tokens are on the board in the next scene.