Validates that answer selected is correct or incorrect
Should probably delegate this logic to another class, but the RuleController will check if the answer is correct now.
NOTE: will only work if you run the game from the first scene, otherwise there is no rule controller present for the task to be delegated to and it wont be able to check if the answer is correct or not.