Fixed the problem with questions not loading when game is run from Build.
When Unity compiles, it only includes the files that it needs to build scenes plus everything it finds in the Assets/Resources folder. So, I've moved the csv to the Resources folder and loaded it as a TextAsset using Resources.Load().
Can't call Load() from a constructor in a monobehavior so moved the instantiation of the QuestionDatabase to the Start() method of the Rule Controller.